Renamed subpanel_layout to subobject_layout
